Refactoring controller libs

+++ dieselController/Inc/journal.h
@@ -12,10 +12,8 @@
 #include "RTC.h"
 #include "storage.h"
 /*------------------------ Define --------------------------------------*/
-#define  LOG_TIME_ERROR   ( ( uint32_t ) 1U )
-
+#define  LOG_TIME_ERROR   1U
 #define  LOG_START_YEAR   2020U
-
 #define  LOG_DAY_MASK     0xF8000000U
 #define  LOG_DAY_SHIFT    27U
 #define  LOG_MONTH_MASK   0x07800000U
